Partially rewrite UI to be native Android
The old HTML UI was full of hacks (like running a private HTTP
server on port 3000) and wasn't really integrated with the rest
of the apps.

This commit begins the rewrite to use native Android widgets,
starting from the assistant (which is somewhat easier, because
it has well defined entry-points we can hook as JavaAPI/JavascriptAPI).
The "My Stuff" page is still in HTML for now, it will be addressed
in a later commit.