Remove assistant history
Replaying the history on resume means that we get multiple messages
because the old ones are not destroyed on pause.
We could replay the history on start but it's not really worth it,
because the history is not on disk so it handles activity stop
but not service stop.
And in any case that's the wrong layer for history.