Rework assistant dispatching mechanism
The sabrina code assumes it can just call the various conversation
delegate methods at any time, but we need to do different things
at different times, including recording messages in the history
or showing a popup notification.
This will be handled by moving as much middle man as possible to
java, and having AssistantDispatcher hold the service side of
the logic, and AssistantFragment the activity side.