Add debug log instrumentation, and upload logs to the server
Some builds (notably release builds, and not debug builds) fail
at startup because they don't find some files we need.
Let's catch that, log that, and upload all the logs to the server
where we can process them later.
Sorry if that's an invasion of privacy, I need to fix this.