Fix hangs caused by a race condition in ControlChannel
If 2 replies happened on the control channel in rapid succession,
it was possible for a reply to be left in partial message and never
seen. The subsequent reader would be left blocking on a read()
call that would never complete until the background thread has something else
to say for whatever reason.