Add a never mind button next to the voice button
Reset is too technical of a word, and the button is too out of the way
at the top. Instead we should offer an easy to access button next
to the text box, to never mind without clearing the history.
The button is only visible when sabrina is in the middle of a dialog,
and is shaped like a "start over" arrow.

Clearing the history is still available from the menu, but that leaves
space for Train in the action bar.