Update launcher code to support node v8.x
Luckily, not many API changes happened. A few changes occurred in
the area of v8::Platform (which is responsible for v8 background
tasks and promises), but we can keep using the older code because
we don't care about shutting down the node loop.

Also, we removed the shared library from the git repository.
They are too big to store in Git, they should be built on demand.