Radically simplify how nodejs is launched
Instead of diving deep into nodejs internals for questionable
reasons (like being able to drive the main loop by hand), which
leads to hangs and weird failures every time we update node,
just use the public node::Start() API. This API essentially runs
a full node process, but without spawning a separate process
(which is not allowed on Android). Communication with Java and
marshalling is still the same, because it happens in a regular
node native module.